server/requirements.txt
fastapi
python-multipart
openai-whisper
uvicorn
ffmpeg-python
